+++ Assets/Code/Common/Physics/Internal/KinematicRigidbody2D.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9cc78c8604ec91b458a4f10477a5a040
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
